+++ Tests/MapboxMapsTests/Foundation/Mocks/MockDisplayLink.swift
@@ -9,6 +9,21 @@ final class MockDisplayLink: DisplayLinkProtocol {
 
     var preferredFramesPerSecond: Int = 0
 
+    // Checking Swift version as a proxy for iOS SDK version to enable
+    // building with iOS SDKs < 15
+    #if swift(>=5.5)
+    @available(iOS 15.0, *)
+    var preferredFrameRateRange: CAFrameRateRange {
+        get {
+            return (_untypedPreferredFrameRateRange as? CAFrameRateRange) ?? .default
+        } set {
+            _untypedPreferredFrameRateRange = newValue
+        }
+    }
+
+    private var _untypedPreferredFrameRateRange: Any?
+    #endif
+
     struct AddParams {
         var runloop: RunLoop
         var mode: RunLoop.Mode